[#578] made commsnode work in local as a separate mode
